1d tensor for orthogonal bsplines added; examples provided
Former-commit-id: f7369d7c6cd697befeee75c195b873f447cb1dcf

+++ combigrid/src/sgpp/combigrid/utils/AnalyticModels.hpp
@@ -169,7 +169,7 @@ struct Genz {
     double ans = 2.0 * M_PI * w;
     size_t numDims = x.size();
     for (size_t k = 0; k < numDims; k++) {
-      ans += 4.5 * (k + 0.5) / static_cast<double>(numDims) * x[k];
+      ans += 4.5 * (static_cast<double>(k) + 0.5) / static_cast<double>(numDims) * x[k];
     }
     return std::cos(ans);
   }
